tts/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistant-tts-worker"
version = "0.1.0"
dependencies = [
    "piper-tts"
]

[tool.setuptools]
py-modules = ["worker", "app"]
